main navigation system is mostly done
