+++ app/(tabs)/(drawer)/terms.tsx
@@ -1,19 +1,16 @@
-import { Link, Stack } from "expo-router";
+import { Link } from "expo-router";
 import { Button, StyleSheet, Text, View } from "react-native";
 
-export default function Conduct() {
+export default function Terms() {
     return (
-      <>
-      <Stack.Screen options={{headerShown: true, title: 'Codes of Conduct'}} />
       <View style={styles.container}>
       <View style={styles.main}>
-      <Text style={styles.title}>Codes of Conduct</Text>
+      <Text style={styles.title}>Terms and Conditions</Text>
         <Link href={'/(tabs)/home'} asChild>
           <Button title="Go to Tabs"/>
         </Link>
       </View>
     </View>
-      </>
       );
   }
