+++ app/_layout.tsx
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { useColorScheme } from "react-native";
+import { useColorScheme, View, Image } from "react-native";
 import { DarkTheme, DefaultTheme, ThemeProvider } from "@react-navigation/native";
 import { TamaguiProvider } from "tamagui";
 import { config } from "./../tamagui.config";
@@ -27,6 +27,31 @@ export {
 } from "expo-router";
 import "../tamagui.css";
 import "@tamagui/core/reset.css";
+import CustomDrawerContent from "./(tabs)/drawer_trigger";
+import Drawer from "expo-router/drawer";
+import { DrawerToggleButton } from "@react-navigation/drawer";
+
+function LogoTitle() {
+  const theme = useColorScheme();
+  const isDarkTheme = theme === "dark";
+  const iconTheme = isDarkTheme ?
+      require('../assets/dark.png') :
+      require('../assets/light.png');
+  return (
+      <View>
+          <Image
+              style={{
+                  alignSelf: 'flex-end',
+                  marginRight: 24,
+                  marginTop: 7,
+                  width: 101,
+                  height: 48
+              }}
+              source={iconTheme}
+          />
+      </View>
+  );
+}
 
 export default function App() {
   const [fontsLoaded, fontError] = useFonts({
@@ -42,6 +67,10 @@ export default function App() {
     OpenSans_700Bold_Italic,
   });
   const colorScheme = useColorScheme();
+  const theme = useColorScheme();
+  const isDarkTheme = theme === "dark";
+  const toodleColor = isDarkTheme ? 'white' : 'black';
+  const barColor = isDarkTheme ? '#140233' : '#FFFFFF'
   useEffect(() => {
     if (fontsLoaded) {
       // can hide splash screen here
@@ -55,11 +84,23 @@ export default function App() {
     <TamaguiProvider config={config} defaultTheme={colorScheme as any}>
       <ThemeProvider value={colorScheme === "dark" ? DarkTheme : DefaultTheme}>
         {Constants.expoConfig?.extra?.storybookEnabled ? <Storybook /> :
-        <Stack  screenOptions={{
-          headerShown: false,
-          }}>
-          <Stack.Screen name="index" options={{ headerShown: false }}/>
-          </Stack>}
+          <Drawer drawerContent={(props) => <CustomDrawerContent
+            name={"Yoongi's Wife"}
+            userName={"tangerines4life"}
+            followers={208}
+            following={67} {...props} />}
+            screenOptions={{
+              headerShown: false,
+              headerLeft: () => <DrawerToggleButton
+                tintColor={toodleColor}
+              />,
+              headerRight: () => <LogoTitle />,
+              headerStyle: {
+                backgroundColor: barColor
+              },
+            }}>
+            <Drawer.Screen name="(tabs)" options={{ title: 'Tabs' }} />
+          </Drawer>}
       </ThemeProvider>
     </TamaguiProvider>
   );
